% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expertscore.R
\name{expertscore}
\alias{expertscore}
\title{Expert score}
\usage{
expertscore(expertmap, probsurface, studyarea)
}
\arguments{
\item{expertmap}{THis is a map}

\item{probsurface}{This is a raster}

\item{studyarea}{This is study area}
}
\value{
This function returns an expert score
}
\description{
This is a short description
}
\details{
Here are a bunch of details about this function
}
\examples{
# load data
data(myexpert1)
data(myprobsurface)
data("mystudyarea")
data("myoccurrences")

# compute expert score
expertscore(expertmap = myexpert1, probsurface = myprobsurface, studyarea = mystudyarea)
plot(myprobsurface)
plot(mystudyarea, add=T)
plot(myoccurrences, add=T)
}
\references{
Mainali et al Conservation Biology
}
\author{
Kumar Mainali, Elie G, T He.
}
