% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{myoccurrences}
\alias{myoccurrences}
\title{Occurence data}
\format{SpatialPointsDataFrame
\describe{
\item{id}{ID of animal}
\item{start}{Date of beginning of collaring}
\item{end}{Date of death or censoring}
\item{fate}{One of "dead", or "censored"}
}}
\source{
This is my source
}
\usage{
data(occurences)
}
\description{
A bunch of occurrence data
}
\examples{
data(occurences)


}
\keyword{data}
