% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{mystudyarea}
\alias{mystudyarea}
\title{Study area}
\format{SpatialPolygonsDataFrame
\describe{
  \item{projection}{"+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +units=m +no_defs"}
}}
\source{
Hypothetical data
}
\usage{
data(mystudyarea)
}
\description{
Union of all expert maps and convex hull of occurrences (this is one of the methods)
}
\examples{
data(mystudyarea)

}
\keyword{data}
