% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{r_memory}
\alias{r_memory}
\title{Raster memory}
\format{An object of class \code{RasterLayer} of dimension 6513 x 4776 x 1.}
\usage{
data(r_memory)
plot(r_memory)
}
\description{
#' test
}
\keyword{datasets}
