% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expertscore.R
\name{expertscore}
\alias{expertscore}
\title{Expert score}
\usage{
expertscore(expertmap, probsurface, studyarea)
}
\arguments{
\item{expertmap}{THis is a map}

\item{probsurface}{This is a raster}

\item{studyarea}{This is study area}
}
\value{
This function returns an expert score
}
\description{
This is a short description
}
\details{
Here are a bunch of details about this function
}
\examples{
# load the hypothetical predicted continuous probability surface
data(myprobsurface)

# load three hypothetical expert maps
data(myexpert1)
data(myexpert2)
data(myexpert3)

# load the hypothetical occurrences
data(myoccurrences)

# load study area
data(mystudyarea)

# plot the data
image(myprobsurface, zlim=c(0,1),col=rev(terrain.colors(100)),xaxt="n",yaxt="n",bty="n",xlab=" ",ylab=" ")
points(myoccurrences, pch=21, col="purple", cex=1)
plot(myexpert1, add=TRUE, border="orange", lwd=2)
plot(myexpert2, add=TRUE, border="red", lwd=2)
plot(myexpert3, add=TRUE, border="cyan", lwd=2)
plot(mystudyarea, add=TRUE, border='blue', lwd=1)

# compute expert score of the three hypothetical expert maps
expertscore(expertmap = myexpert1, probsurface = myprobsurface, studyarea = mystudyarea)
expertscore(expertmap = myexpert2, probsurface = myprobsurface, studyarea = mystudyarea)
expertscore(expertmap = myexpert3, probsurface = myprobsurface, studyarea = mystudyarea)
}
\references{
Mainali et al Conservation Biology
}
\author{
Kumar Mainali, Elie G, T He.
}
